+++ lib/end_of_life/repository.rb
@@ -1,5 +1,3 @@
-require "dry-monads"
-
 module EndOfLife
   class Repository
     class << self