+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ### Added
 
+- Fetch EOL Ruby versions from endoflife.date API. This ensures we always use up-to-date data but keep the embedded JSON as a fallback.
+
 - Allow users to specify the maximum number of days away a version can be from EOL. It defaults to 0.
 
 ```sh